Optimize Content-Length: 0 handling
Create and use a pre-encoded HttpFields.CONTENT_LENGTH_0 constant
Used the constant in more places
renamed the `putLongField` methods of HttpFields.Mutable to just `put`

+++ jetty-core/jetty-http/src/main/java/org/eclipse/jetty/http/HttpField.java
@@ -501,7 +501,7 @@ public class HttpField
         @Override
         public int getIntValue()
         {
-            return (int)_long;
+            return Math.toIntExact(_long);
         }
 
         @Override
